Add parsing methods for legistar calendar rss
This is needed for the sole purpose of gathering the "Agenda Packet"
pdf files.
The packets are not linked anywhere in Granicus, so we have to
scrape the calendar rss feed from Legistar and match items
to clips by almost guessing (since their RSS feed is lacking
most usable info).

Then, from the `<link>` value of matched RSS items, we can
parse the `MeetingDetail` page and get the agenda packet url.

That's a LOT just for one item, but it's important so...