Use formatted datetime from granicus template output
The timestamp value that it outputs is off by a day for some
clips apparently.  This was missed previously since it was
such a low ratio.
The formatted datetime string is consistent though, so we'll use
that instead.
Hey, who needs consistency right?  smh

I had to manually find the clips that were wrong and remove/rename
them.  Anyone else that uses this in the future won't have to
suffer my fate.  You're welcome!