Limit concurrent parsing of detail pages
Another rant...
Apparently they rate limit that page or something.  What's
worse is that the status code for the response is 200, but the
error information is happily "displayed" in the html body.

At least they put "Error" in the `<title>` tags though! Thanks
for that!
Anyway, we can't load up more than a couple at a time before it
starts to fail, so let's just do one at a time (sloooowwwww)
</rant>