Prevent item duplication when its GUID changes
Yes, I did just say that... An item's guid ACTUALLY CHANGES
whenever it's updated in legistar!

The guid field of the "RSS" feed has datetime information appended
to it for whatever reason like this:
`2024-09-24-18-52-26`

That part of the "guid" changes, but it still refers to the same item
which defeats the ENTIRE PURPOSE of a Globally Unique IDentifier!

Strip the pointless junk off the end and use the real guid to
ensure uniqueness.

</rant>