Add check for unknown download error
This is a weird one. Some pdf files after downloading
have a local filesize of exactly 1245 bytes.
The `Content-Length` in the response headers are much
larger and seem to always match when downloading manually.

Not sure if this is a bug on the server side, aiohttp or
in this codebase.